requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-seed==0.2.2
djangorestframework==3.11.1
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
